// src/app/model/flocash/response/flocash-order.response.ts
export class FlocashOrderResponse {
    amount: number;
    capturedAmount: number;
    refundedAmount: number;
    orderDate: number;
    currency: string;
    currencyName: string;
    custom: string;
    item_name: string;
    item_price: string;
    quantity: string;
    orderId: string;
    tracking: string;
    traceNumber: string;
    partnerMessage: string;
    status: any;
    statusDesc: string;
    paymentChannel: string;
    payer: any;
    payOption: any;
    approveCode: string;
    redirect: any;

}